+++ TelegramGroupsAdmin.ContentDetection/Models/TrainingSample.cs
@@ -27,7 +27,7 @@ public record TrainingSample
     /// <summary>
     /// Original message ID (for traceability and debugging).
     /// </summary>
-    public required long MessageId { get; init; }
+    public required int MessageId { get; init; }
 
     /// <summary>
     /// User ID who labeled this sample (null for implicit/auto-detected samples).
