+++ TelegramGroupsAdmin.Telegram/Models/MessageEditRecord.cs
@@ -5,7 +5,7 @@ namespace TelegramGroupsAdmin.Telegram.Models;
 /// </summary>
 public record MessageEditRecord(
     long Id,
-    long MessageId,
+    int MessageId,
     string? OldText,
     string? NewText,
     DateTimeOffset EditDate,
